flashcards/requirements-dev.txt
-r requirements.txt

django-debug-toolbar==5.1.0
pytest==8.3.5
pylint==3.3.6
pylint-django==2.6.1
python-dotenv==1.1.0
